+++ sdk/ai/azure-ai-generative/azure/ai/generative/evaluate/_evaluation_result.py
@@ -1,5 +1,8 @@
 from typing import Dict
 
+from azure.ai.generative.evaluate._utils import _get_ai_studio_url
+
+
 class EvaluationResult(object):
 
     def __init__(self, metrics_summary: Dict[str, float], artifacts: Dict[str, str], **kwargs):
@@ -7,6 +10,8 @@ class EvaluationResult(object):
         self._artifacts = artifacts
         self._tracking_uri = kwargs.get("tracking_uri")
         self._evaluation_id = kwargs.get("evaluation_id")
+        if self._tracking_uri:
+            self._studio_url = _get_ai_studio_url(self._tracking_uri, self._evaluation_id)
 
     @property
     def metrics_summary(self) -> Dict[str, float]:
@@ -19,6 +24,10 @@ class EvaluationResult(object):
     @property
     def tracking_uri(self) -> str:
         return self._tracking_uri
+
+    @property
+    def studio_url(self) -> str:
+        return self._studio_url
 
     def download_evaluation_artifacts(self, path: str) -> str:
         from mlflow.artifacts import download_artifacts