+++ sdk/appconfiguration/azure-appconfiguration/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking Changes
 
 ### Bugs Fixed
+- Removed keyword argument `etag` in `set_read_only()` as it's not in use.
 
 ### Other Changes
 - Supported datetime type for keyword argument `accept_datetime` in `get_snapshot_configuration_settings()`, `list_snapshot_configuration_settings()` and `list_revisions()`.